refactor(cli): extract final leaves into cohesive command modules (5.0-alpha)
Final cut of the cli.ts god-module split. Every remaining standalone leaf
handler moves into a cohesive commands/* module; cli.ts becomes a thin
dispatcher (imports + COMMANDS/COMMAND_TIERS/COMMAND_HELP + main + cmdHelp).

New modules (each handler exported; helpers stay module-private):
- commands/maintain.ts — cmdHeal, cmdPkg (remediation + triage-gated install)
- commands/agent.ts — cmdSkills, cmdEscalate, cmdAgentConfig, cmdGovernance,
  cmdStatusline, cmdTeam (agent-facing + governance)
- commands/coverage.ts — cmdSelfAudit, cmdCoverage, cmdAnalyze (+ private
  aggregateAdvisories) (coverage/audit/analysis reporting)
- commands/gate.ts — cmdGateBash, cmdGateEnv (+ private gateFormat) (PreToolUse
  hook gates)

cli.ts imports them back; COMMANDS table unchanged. Pruned all now-unused
imports.

Verification: tsc clean, eslint 0 errors, full suite 2580 passing, prettier
clean, runtime dispatch smoke-tested (heal/skills/governance/self-audit/
coverage/team/pkg/statusline/gate-bash).

cli.ts: 1641 -> 717 lines. The migration is complete: 6330 -> 717 (~89%
smaller); ~40 handlers now live in single-concern commands/* modules behind an
unchanged COMMANDS dispatch table.
