+++ src/cli.ts
@@ -2,31 +2,12 @@
 
 import { readFileSync, realpathSync } from "node:fs";
 import { fileURLToPath } from "node:url";
-import { resolve, dirname, join } from "node:path";
+import { dirname, join } from "node:path";
 import { loadConfig } from "./config.js";
-import { checkTools } from "./check-tools.js";
-import { checkServices } from "./check-services.js";
-import { checkSecrets } from "./check-secrets.js";
-import { checkSecurity, type SecurityCheckResult } from "./check-security.js";
-import { hasFlag, flagValue } from "./utils/flags.js";
-import { analyzeRepo, renderClaudeMd, renderRulesMd } from "./analyze.js";
-import { scanTranscripts } from "./scan-transcripts.js";
+import { hasFlag } from "./utils/flags.js";
 import { generateCompletions } from "./completions.js";
 import { checkForUpdate, printUpdateNotice } from "./update-check.js";
-import { checkSkills } from "./check-skills.js";
-import { collectEscalations, formatEscalationMessage } from "./escalate.js";
-import { checkRevocationStatus } from "./revocation.js";
-import { getBudgetStatus, formatBudgetStatus } from "./budget.js";
-import { formatGovernanceStatus, mergeGovernanceConfigAsync } from "./governance.js";
-import { withGovernance } from "./governance-middleware.js";
 import { SKIPPED_COMMITS_LOG } from "./hooks.js";
-import {
-  writeAgentConfig,
-  detectAgentTargets,
-  installKitPermissions,
-  installAllInstallGates,
-  installAiderRules,
-} from "./agent-config.js";
 import { cmdFix } from "./fix.js";
 import { c } from "./utils/colors.js";
 import { KIT_FILE, resolveConfigPath } from "./cli-shared.js";
@@ -56,6 +37,17 @@ import {
   cmdVersion,
 } from "./commands/info.js";
 import { cmdCreatePlugin, cmdClone, cmdRun, cmdOpen } from "./commands/project.js";
+import { cmdHeal, cmdPkg } from "./commands/maintain.js";
+import {
+  cmdSkills,
+  cmdEscalate,
+  cmdAgentConfig,
+  cmdGovernance,
+  cmdStatusline,
+  cmdTeam,
+} from "./commands/agent.js";
+import { cmdSelfAudit, cmdCoverage, cmdAnalyze } from "./commands/coverage.js";
+import { cmdGateBash, cmdGateEnv } from "./commands/gate.js";
 import { cmdCi } from "./commands/ci.js";
 import { cmdSentinel } from "./commands/sentinel.js";
 import { cmdStandards, cmdBaseline } from "./commands/standards.js";
@@ -63,14 +55,6 @@ import { cmdCheck } from "./commands/check.js";
 import { cmdDesign } from "./commands/design.js";
 import { cmdInstall } from "./commands/install.js";
 import { cmdLogin } from "./commands/login.js";
-import {
-  type CiFormat,
-  type JsonCheck,
-  type JsonCheckOutput,
-  detectCiFormat,
-  emitGithubAnnotations,
-  emitGitlabJunit,
-} from "./cli-checks-shared.js";
 import { cmdAuth } from "./commands/auth.js";
 import { cmdAudit } from "./commands/audit.js";
 import { cmdMcp } from "./commands/mcp.js";
@@ -78,13 +62,7 @@ import { cmdHooks } from "./commands/hooks.js";
 import { cmdPlugin } from "./plugins-cli.js";
 import { cmdTriage } from "./commands/triage.js";
 import { cmdSlopsquat } from "./commands/slopsquat.js";
-import { parsePkgSpec, installPkg } from "./pkg.js";
 import { cmdMemory } from "./commands/memory.js";
-import { resolveKitRoot, runSelfAudit, SELF_AUDIT_RULES } from "./self-audit.js";
-import { buildCoverageReport, formatCoverageText, type Bucket } from "./coverage/coverage.js";
-import { buildStandardReport, formatStandardText } from "./coverage/standard.js";
-import { OWASP_LLM_TOP10 } from "./coverage/owasp-llm-top10.js";
-import { SSDF_218A } from "./coverage/ssdf-218a.js";
 import { escapeWorkflowCmd, xmlEscape } from "./utils/ci-escape.js";
 
 // Re-exported for tests + downstream emitters (ci-escaping.test.ts imports these
@@ -100,641 +78,6 @@ const __dirname = dirname(fileURLToPath(import.meta.url));
 const KIT_VERSION = (
   JSON.parse(readFileSync(join(__dirname, "..", "package.json"), "utf-8")) as { version: string }
 ).version;
-
-/** Map a security finding to a short, actionable PAL title/detail. */
-async function cmdHeal(): Promise<boolean> {
-  const dryRun = hasFlag(process.argv, "--dry-run");
-  const agent = hasFlag(process.argv, "--agent");
-  console.log(
-    `${c.bold}${c.cyan}kit heal${c.reset}${dryRun ? `${c.dim} (dry-run)${c.reset}` : ""}`,
-  );
-  console.log(`${c.dim}${"─".repeat(50)}${c.reset}\n`);
-
-  const { runHeal } = await import("./heal.js");
-  // Progress goes to stderr: live feedback for a human watching, without
-  // polluting the machine-readable proposals on stdout (--agent).
-  const res = await runHeal({ dryRun, onProgress: (m) => console.error(`${c.dim}${m}${c.reset}`) });
-  console.log();
-
-  if (dryRun) {
-    if (res.plannedSafe.length > 0) {
-      console.log(`${c.bold}Would auto-fix (safe):${c.reset}`);
-      for (const k of res.plannedSafe) console.log(`  ${c.green}✓${c.reset} ${k}`);
-    } else {
-      console.log(`${c.dim}Nothing to auto-fix.${c.reset}`);
-    }
-  } else if (res.healed.length > 0) {
-    console.log(`${c.green}${c.bold}Healed ${res.healed.length}:${c.reset}`);
-    for (const k of res.healed) console.log(`  ${c.green}✓${c.reset} ${k}`);
-  }
-
-  // FAIL-CLOSED — loud, never auto-healed.
-  if (res.failClosed.length > 0) {
-    console.log(
-      `\n${c.red}${c.bold}⚠ FAIL-CLOSED — not auto-healed (possible tampering):${c.reset}`,
-    );
-    for (const r of res.failClosed) {
-      console.log(`  ${c.red}✗${c.reset} ${r.name}: ${r.detail}`);
-      if (r.suggestion) console.log(`    ${c.dim}${r.suggestion}${c.reset}`);
-    }
-  }
-
-  // GATED — proposed, never auto-run by kit.
-  if (res.gated.length > 0) {
-    console.log(`\n${c.yellow}${c.bold}Gated — needs you (kit won't auto-run these):${c.reset}`);
-    for (const g of res.gated) {
-      console.log(`  ${c.yellow}!${c.reset} ${g.name}: ${g.issue}`);
-      console.log(`    ${c.dim}→ ${g.action}${c.reset}`);
-    }
-    if (agent) {
-      console.log(
-        `\n${c.dim}# agent: each command below still hits the elevation gate + audit log${c.reset}`,
-      );
-      for (const g of res.gated) console.log(g.action);
-    }
-  }
-
-  const green = res.failClosed.length === 0 && res.gated.length === 0;
-  console.log();
-  if (!dryRun) {
-    console.log(
-      green
-        ? `${c.green}${c.bold}All findings healed or clean ✓${c.reset}`
-        : `${c.yellow}Auto-heal done; items above need you.${c.reset}`,
-    );
-  }
-  return green;
-}
-
-async function cmdSkills(): Promise<boolean> {
-  const config = await loadConfig(resolveConfigPath());
-
-  if (!config.skills) {
-    console.log(`${c.dim}No skills configured in ${KIT_FILE}${c.reset}`);
-    return true;
-  }
-
-  const registry = config.skills.registry ?? "clawhub";
-  console.log(`${c.bold}${c.cyan}Skills${c.reset}  ${c.dim}(registry: ${registry})${c.reset}\n`);
-
-  const results = await checkSkills(config.skills);
-
-  if (results.length === 0) {
-    console.log(`${c.dim}No skills listed in ${KIT_FILE}${c.reset}`);
-    return true;
-  }
-
-  const nameWidth = Math.max(12, ...results.map((r) => r.name.length)) + 2;
-
-  for (const r of results) {
-    const icon = r.installed
-      ? `${c.green}✓${c.reset}`
-      : r.required
-        ? `${c.red}✗${c.reset}`
-        : `${c.yellow}!${c.reset}`;
-    const name = r.name + " ".repeat(Math.max(0, nameWidth - r.name.length));
-    const tag = r.required ? `${c.dim}[required]${c.reset}` : `${c.dim}[optional]${c.reset}`;
-    const status = r.installed ? `${c.green}installed${c.reset}` : `${c.red}missing${c.reset}`;
-    console.log(`  ${icon} ${name} ${tag} ${status}  ${c.dim}${r.versionSpec}${c.reset}`);
-  }
-
-  const missing = results.filter((r) => !r.installed);
-  if (missing.length > 0) {
-    console.log();
-    console.log(`${c.bold}To install missing skills:${c.reset}`);
-    for (const m of missing) {
-      console.log(`  ${c.cyan}openclaw install ${registry}/${m.name}${c.reset}`);
-    }
-  }
-
-  console.log();
-  const requiredMissing = results.filter((r) => r.required && !r.installed);
-  return requiredMissing.length === 0;
-}
-
-async function cmdEscalate(): Promise<boolean> {
-  const config = await loadConfig(resolveConfigPath());
-
-  return await withGovernance(
-    config,
-    {
-      operation: "escalate",
-      operationType: "read",
-      metadata: {},
-    },
-    async () => {
-      const toolResults = config.tools ? await checkTools(config.tools) : [];
-      const serviceResults = config.services ? await checkServices(config.services) : [];
-      const secretResults = config.secrets
-        ? await checkSecrets(config.secrets)
-        : { templateExists: null, keys: [] };
-
-      const items = collectEscalations(toolResults, serviceResults, secretResults.keys);
-
-      if (items.length === 0) {
-        console.log(`${c.green}All checks passed — nothing to escalate.${c.reset}`);
-        return true;
-      }
-
-      const message = formatEscalationMessage(items, process.cwd());
-      console.log(`${c.bold}${c.yellow}Escalation summary${c.reset}\n`);
-      console.log(message);
-
-      // Write to file for easy copy/paste or piping
-      const { writeFile } = await import("node:fs/promises");
-      const escalationPath = resolve(process.cwd(), ".kit-escalation.txt");
-      await writeFile(escalationPath, message, "utf-8");
-      console.log(`${c.dim}Written to ${escalationPath}${c.reset}`);
-      console.log(`${c.dim}Send this to the project owner for manual resolution.${c.reset}\n`);
-
-      return false;
-    },
-  );
-}
-
-async function cmdAgentConfig(): Promise<boolean> {
-  console.log(`${c.bold}${c.cyan}kit agent-config${c.reset}`);
-  console.log(`${c.dim}${"─".repeat(50)}${c.reset}\n`);
-
-  const targets = detectAgentTargets();
-  console.log(
-    `${c.dim}Teaching ${targets.map((t) => `${c.reset}${c.bold}${t.agent}${c.reset}${c.dim}`).join(", ")} to use kit ` +
-      `(managed block in their rules file).${c.reset}\n`,
-  );
-
-  const results = await writeAgentConfig();
-  // Aider needs a bespoke installer (CONVENTIONS.md + a `read:` entry in
-  // .aider.conf.yml — it auto-reads no rules file), so it's not an AGENT_TARGETS row.
-  const aider = await installAiderRules();
-  if (aider.detail !== "no Aider project detected") results.push(aider);
-  let failed = false;
-  for (const r of results) {
-    if (r.action === "failed") {
-      failed = true;
-      console.log(`  ${c.red}✗${c.reset} ${r.agent} (${r.file}): ${r.detail}`);
-    } else {
-      const mark = r.action === "unchanged" ? `${c.dim}=` : `${c.green}✓`;
-      console.log(`  ${mark}${c.reset} ${r.agent} ${c.dim}→ ${r.file} (${r.action})${c.reset}`);
-    }
-  }
-  // Let the agent actually RUN kit: grant read-only kit commands in
-  // .claude/settings.json, so they don't hit the permission wall in auto mode.
-  const perms = await installKitPermissions();
-  if (perms.action === "created" || perms.action === "updated") {
-    console.log(
-      `\n  ${c.green}✓${c.reset} allowed ${c.bold}${perms.added.length}${c.reset} read-only kit command(s) in ${c.dim}${perms.file}${c.reset} ${c.dim}(so the agent can run them without a prompt)${c.reset}`,
-    );
-  } else if (perms.action === "unchanged") {
-    console.log(`\n  ${c.dim}= read-only kit commands already allowed in ${perms.file}${c.reset}`);
-  } else if (perms.action === "failed") {
-    console.log(`\n  ${c.yellow}!${c.reset} could not update ${perms.file}: ${perms.detail}`);
-  }
-  // Default-ON: the true blocking gates (PreToolUse hooks) — un-triaged installs
-  // and plaintext secrets aimed at .env* are blocked BEFORE they run/land. The
-  // rules block above only advises; these enforce. An instruction the agent can
-  // ignore is a false green, so enforcement is the default; --no-install-gate
-  // opts out (--install-gate is still accepted for backward compatibility).
-  if (!hasFlag(process.argv, "--no-install-gate")) {
-    console.log(
-      `\n  ${c.bold}PreToolUse gates${c.reset} ${c.dim}(block un-triaged installs + plaintext .env* secrets before they happen; --no-install-gate to skip):${c.reset}`,
-    );
-    for (const { agent, result } of await installAllInstallGates()) {
-      if (result.action === "created" || result.action === "updated") {
-        console.log(
-          `    ${c.green}✓${c.reset} ${agent} ${c.dim}→ ${result.file}${result.detail ? ` (${result.detail})` : ""}${c.reset}`,
-        );
-      } else if (result.action === "unchanged") {
-        console.log(`    ${c.dim}= ${agent} already wired (${result.file})${c.reset}`);
-      } else {
-        console.log(`    ${c.dim}· ${agent} skipped: ${result.detail ?? result.action}${c.reset}`);
-      }
-    }
-  }
-  console.log(
-    `\n${c.dim}Blocks regenerate in place on re-run; edit outside the markers freely. ` +
-      `Mutating kit commands (secrets/fix/hooks) still prompt by design.${c.reset}`,
-  );
-  console.log(
-    `\n${c.bold}Agent support${c.reset} ${c.dim}(what kit wires up per agent):${c.reset}\n` +
-      `  ${c.dim}· Memory index: Claude Code, Codex, Cursor, Cline, Gemini, Continue, Amazon Q, Kiro, Factory Droid, Aider, Antigravity, OpenCode${c.reset}\n` +
-      `  ${c.dim}· "use kit" rules block: Claude Code, Codex, Cursor, Cline, OpenCode${c.reset}\n` +
-      `  ${c.dim}· Config/secret audit (kit agent-audit): Claude Code, Codex, Cursor, OpenCode (+ generic .mcp.json)${c.reset}\n` +
-      `  ${c.dim}· Permission allowlist + auto-capture hooks: Claude Code${c.reset}\n` +
-      `  ${c.dim}· Blocking install-gate: Claude Code, Codex, Amazon Q, Kiro, Factory Droid, Augment, Antigravity, Gemini CLI, Cursor (hooks), OpenCode (plugin), Cline (PreToolUse shim); Continue has no gate surface (#146)${c.reset}\n` +
-      `  ${c.dim}The agent-agnostic enforcement floor is git hooks (${c.reset}${c.bold}kit hooks${c.reset}${c.dim}); the rules block only advises.${c.reset}`,
-  );
-  return !failed;
-}
-
-async function cmdGovernance(): Promise<boolean> {
-  const config = await loadConfig(resolveConfigPath());
-  const governanceConfig = await mergeGovernanceConfigAsync(config.governance);
-
-  console.log(`${c.bold}${c.cyan}Governance Status${c.reset}`);
-  console.log(`${c.dim}${"─".repeat(50)}${c.reset}\n`);
-
-  // Display governance configuration
-  console.log(formatGovernanceStatus(governanceConfig));
-  console.log();
-
-  // Check revocation status
-  if (governanceConfig.revocation.enabled) {
-    const revoked = await checkRevocationStatus(config.governance);
-    const icon = revoked ? `${c.red}✗${c.reset}` : `${c.green}✓${c.reset}`;
-    const status = revoked ? `${c.red}REVOKED${c.reset}` : `${c.green}Active${c.reset}`;
-    console.log(`  ${icon} Access Status: ${status}`);
-  }
-
-  // Display budget status
-  if (governanceConfig.agent.max_tokens_per_day || governanceConfig.agent.max_operations_per_hour) {
-    console.log();
-    const budgetStatus = await getBudgetStatus(config.governance);
-    console.log(formatBudgetStatus(budgetStatus));
-  }
-
-  // Display agent info
-  if (governanceConfig.agent.id || governanceConfig.agent.name) {
-    console.log();
-    console.log("Agent Information");
-    console.log("─".repeat(50));
-    if (governanceConfig.agent.name) {
-      console.log(`Name: ${governanceConfig.agent.name}`);
-    }
-    if (governanceConfig.agent.id) {
-      console.log(`ID: ${governanceConfig.agent.id}`);
-    }
-  }
-
-  console.log();
-  return true;
-}
-
-/**
- * `kit statusline` — one compact, fast, read-only line for ANY harness's info bar
- * (Claude Code statusLine, a shell PS1, …): setup score for the active mode + an
- * "update available" mark + the open PAL count. Agent-agnostic; never blocks
- * (cached update only, file-presence subsystem checks). Assembly lives in
- * statusline.ts so the memory SessionStart hook injects the SAME line as context.
- */
-async function cmdStatusline(): Promise<boolean> {
-  process.env.KIT_NO_UPDATE_CHECK = "1"; // never let the post-command notice pollute the single line
-  const { buildStatuslineText } = await import("./statusline.js");
-  console.log(await buildStatuslineText({ modeFlag: flagValue(process.argv, "--mode") }));
-  return true;
-}
-
-async function cmdSelfAudit(): Promise<boolean> {
-  const args = process.argv.slice(2);
-
-  // --list-rules: print the registry and exit (no audit run).
-  if (hasFlag(args, "--list-rules")) {
-    for (const r of SELF_AUDIT_RULES) {
-      console.log(
-        `${r.id}\t${r.name}\t${r.detectionClass}\t${r.severity}\t${r.enabled ? "enabled" : "disabled"}`,
-      );
-    }
-    return true;
-  }
-
-  const formatArg = args.find((a) => a.startsWith("--format="))?.split("=")[1] as
-    | CiFormat
-    | undefined;
-  const failOnWarning = hasFlag(args, "--fail-on-warning");
-  const jsonMode = hasFlag(args, "--json");
-  const format: CiFormat = formatArg ?? (jsonMode ? "json" : detectCiFormat());
-  const onlyArg = args.find((a) => a.startsWith("--only="))?.split("=")[1];
-  const only = onlyArg
-    ? onlyArg
-        .split(",")
-        .map((s) => s.trim())
-        .filter(Boolean)
-    : undefined;
-
-  // self-audit targets kit's OWN source tree — not the user's project. Locate it
-  // by walking up to the sandstream-kit package.json; if kit is installed in a way
-  // that hides its sources, skip cleanly (not a failure).
-  const root = resolveKitRoot();
-  if (root === null) {
-    console.log(
-      `${c.yellow}kit source tree not found; self-audit targets kit itself and has nothing to scan here (skipped).${c.reset}`,
-    );
-    return true;
-  }
-
-  const results = runSelfAudit(root, only ? { only } : undefined);
-
-  // Map SecurityCheckResult[] -> the CI JsonCheck shape, carrying file:line and
-  // severity through. `info` (severity 'low') marks advisory findings: inventory,
-  // not gating signal. Advisories are counted separately (NOT as warnings) so they
-  // never trip --fail-on-warning, and annotation emitters exclude them.
-  const checks: (JsonCheck & { info: boolean })[] = results.map((r) => ({
-    name: r.name,
-    status: r.status,
-    detail: r.detail,
-    category: r.category,
-    files: r.files,
-    severity: r.severity,
-    info: r.severity === "low",
-  }));
-
-  const summary = checks.reduce(
-    (acc, ch) => {
-      if (ch.status === "pass") acc.passed++;
-      else if (ch.status === "fail") acc.failed++;
-      // Advisory (info) findings are tallied apart from real warnings.
-      else if (ch.status === "warn") {
-        if (ch.info) acc.advisories++;
-        else acc.warnings++;
-      } else acc.skipped++;
-      return acc;
-    },
-    { passed: 0, failed: 0, warnings: 0, skipped: 0, advisories: 0 },
-  );
-
-  // Advisories never gate (not even under --fail-on-warning): only real warnings do.
-  const allOk = summary.failed === 0 && (!failOnWarning || summary.warnings === 0);
-
-  // One aggregated line per advisory class (e.g. "toolchain-pin: 72 ... (advisory)").
-  const advisoryLines = aggregateAdvisories(checks.filter((ch) => ch.info));
-
-  if (format === "github") {
-    // Advisory (info-severity) findings are excluded from CI annotations — they are
-    // inventory, not gating signal — but remain in text/json output.
-    emitGithubAnnotations(checks.filter((ch) => !ch.info));
-    console.log(
-      `kit self-audit: ${summary.passed} passed, ${summary.failed} failed, ${summary.warnings} warnings, ${summary.advisories} advisories`,
-    );
-  } else if (format === "gitlab") {
-    emitGitlabJunit(
-      checks.filter((ch) => !ch.info),
-      allOk,
-    );
-    console.log(
-      `kit self-audit: ${summary.passed} passed, ${summary.failed} failed, ${summary.warnings} warnings, ${summary.advisories} advisories`,
-    );
-  } else if (format === "json") {
-    const output: JsonCheckOutput = {
-      ok: allOk,
-      // Drop the internal `info` marker; keep advisories compact (one row per class).
-      checks: [
-        ...checks.filter((ch) => !ch.info).map(({ info: _info, ...ch }) => ch),
-        ...advisoryLines.map((a) => ({
-          name: a.cls,
-          status: "warn" as const,
-          detail: `${a.count} ${a.label} (advisory)`,
-          category: a.cls,
-          severity: "low" as const,
-        })),
-      ],
-      summary,
-    };
-    console.log(JSON.stringify(output, null, 2));
-  } else {
-    // text — grouped by detection class, PASS/WARN/FAIL per finding; advisories
-    // collapsed to one line per class at the end.
-    console.log(`${c.bold}${c.cyan}kit self-audit${c.reset}`);
-    console.log(`${c.dim}${"─".repeat(50)}${c.reset}`);
-    const gating = checks.filter((ch) => !ch.info);
-    const byClass = new Map<string, typeof checks>();
-    for (const ch of gating) {
-      const cls = ch.category;
-      const bucket = byClass.get(cls);
-      if (bucket) bucket.push(ch);
-      else byClass.set(cls, [ch]);
-    }
-    for (const [cls, group] of byClass) {
-      console.log(`\n${c.bold}${cls}${c.reset}`);
-      for (const ch of group) {
-        const tag =
-          ch.status === "pass"
-            ? `${c.green}PASS${c.reset}`
-            : ch.status === "warn"
-              ? `${c.yellow}WARN${c.reset}`
-              : ch.status === "fail"
-                ? `${c.red}FAIL${c.reset}`
-                : `${c.dim}SKIP${c.reset}`;
-        const where = ch.files && ch.files.length > 0 ? ` ${c.dim}(${ch.files[0]})${c.reset}` : "";
-        console.log(`  ${tag} ${ch.name}: ${ch.detail}${where}`);
-      }
-    }
-    if (advisoryLines.length > 0) {
-      console.log(`\n${c.bold}advisories${c.reset}`);
-      for (const a of advisoryLines) {
-        console.log(`  ${c.dim}${a.cls}: ${a.count} ${a.label}${c.reset}`);
-      }
-    }
-    console.log(
-      `\n${SELF_AUDIT_RULES.length} rules, ${summary.failed} fail, ${summary.warnings} warn, ${summary.advisories} advisory`,
-    );
-  }
-
-  return allOk;
-}
-
-/**
- * `kit coverage` — map kit's deterministic checks to the vendored OWASP ASVS L2
- * subset and report which controls kit auto-verifies vs gap / manual / n/a.
- *
- * This is an EVIDENCE map, not a compliance attestation: it never claims kit
- * makes a project "compliant" or "certified". It is the deterministic evidence
- * source a GRC tool (Vanta, Drata, ...) consumes — not a replacement for one.
- * Output is fully deterministic (the report is pure), so it is safe to diff in CI.
- */
-async function cmdCoverage(): Promise<boolean> {
-  const args = process.argv.slice(2);
-  const formatArg = args.find((a) => a.startsWith("--format="))?.split("=")[1];
-  const jsonMode = hasFlag(args, "--json") || formatArg === "json";
-  const verify = hasFlag(args, "--verify");
-
-  // --verify binds AUTO controls to the ACTUAL latest backing-check results, so
-  // "auto" reads as verified/failing/not-run instead of merely "a check is mapped".
-  // Match is by concrete check/rule name; unmatched backing checks stay "not-run".
-  let results: Awaited<ReturnType<typeof checkSecurity>> | undefined;
-  if (verify) {
-    const security = await checkSecurity();
-    // self-audit only binds when kit's own source tree is reachable (it scans kit,
-    // not the user's project); skip it cleanly otherwise — the security results
-    // still bind. runSelfAudit is only meaningful on kit's own checkout.
-    const root = resolveKitRoot();
-    const selfAudit = root ? runSelfAudit(root) : [];
-    // Command-backed evidence cheap enough to run inline (#206): CI hardening
-    // lint + transcript credential scan. Synthesized under the exact ids the
-    // coverage mapping cites, so those controls bind to a live run instead of
-    // reading not-run. Heavier command evidence (kit secrets validate) stays
-    // honestly unbound.
-    const { runGhaAudit } = await import("./gha-audit.js");
-    const { runCiAudit } = await import("./ci-audit.js");
-    const ciResults = [...runGhaAudit(process.cwd()), ...runCiAudit(process.cwd())];
-    const ciFails = ciResults.filter((r) => r.status === "fail").length;
-    const transcriptHits = await scanTranscripts(process.cwd());
-    const commandEvidence: SecurityCheckResult[] = [
-      {
-        category: "supply-chain",
-        name: "gha-audit",
-        status: ciResults.length === 0 ? "skip" : ciFails > 0 ? "fail" : "pass",
-        detail:
-          ciResults.length === 0
-            ? "no CI workflows to lint"
-            : `${ciResults.length} CI hardening check(s), ${ciFails} failing`,
-      },
-      {
-        category: "secrets",
-        name: "scan-transcripts",
-        status: transcriptHits.length > 0 ? "warn" : "pass",
-        detail:
-          transcriptHits.length > 0
-            ? `${transcriptHits.length} credential-shaped hit(s) in agent transcripts`
-            : "no credentials found in agent transcripts",
-      },
-    ];
-    results = [...security, ...selfAudit, ...commandEvidence];
-  }
-
-  const colorBucket = (bucket: Bucket, label: string): string => {
-    const tint =
-      bucket === "auto"
-        ? c.green
-        : bucket === "gap"
-          ? c.yellow
-          : bucket === "manual"
-            ? c.cyan
-            : c.dim;
-    return `${tint}${label}${c.reset}`;
-  };
-
-  // --standard selects which pinned standard to map against (default ASVS, the
-  // original path). llm-top10 / ssdf route through the generic evidence-map engine.
-  const standard = args.find((a) => a.startsWith("--standard="))?.split("=")[1] ?? "asvs";
-  const STANDARDS = { "llm-top10": OWASP_LLM_TOP10, ssdf: SSDF_218A } as const;
-  if (standard !== "asvs") {
-    const descriptor = STANDARDS[standard as keyof typeof STANDARDS];
-    if (!descriptor) {
-      console.error(
-        `${c.red}unknown --standard '${standard}' (use: asvs | llm-top10 | ssdf)${c.reset}`,
-      );
-      process.exitCode = 1;
-      return false;
-    }
-    const stdReport = buildStandardReport(descriptor, results);
-    console.log(
-      jsonMode ? JSON.stringify(stdReport, null, 2) : formatStandardText(stdReport, colorBucket),
-    );
-    return true;
-  }
-
-  const report = buildCoverageReport(results);
-  if (jsonMode) {
-    console.log(JSON.stringify(report, null, 2));
-    return true;
-  }
-  console.log(formatCoverageText(report, colorBucket));
-  return true;
-}
-
-/**
- * Collapse advisory (info) checks into one aggregated row per detection class.
- * Label is derived from the class slug so the line reads e.g.
- * "self-audit/toolchain-pin: 72 third-party CLI execs (advisory)".
- */
-function aggregateAdvisories(
-  advisories: { category: string }[],
-): { cls: string; count: number; label: string }[] {
-  const ADVISORY_LABELS: Record<string, string> = {
-    "self-audit/toolchain-pin": "third-party CLI execs",
-    "self-audit/env-trust": "env-gated check relaxations",
-  };
-  const byClass = new Map<string, number>();
-  for (const a of advisories) byClass.set(a.category, (byClass.get(a.category) ?? 0) + 1);
-  return [...byClass.entries()].map(([cls, count]) => ({
-    cls,
-    count,
-    label: ADVISORY_LABELS[cls] ?? "advisory findings",
-  }));
-}
-
-async function cmdAnalyze(): Promise<boolean> {
-  // Flags: --claude, --rules to emit one or the other; default is both.
-  // --write <dir> persists the drafts (suffix .draft.md so the user reviews
-  // before committing).
-  const args = process.argv.slice(3);
-  const wantClaude =
-    hasFlag(args, "--claude") || (!hasFlag(args, "--rules") && !hasFlag(args, "--claude"));
-  const wantRules =
-    hasFlag(args, "--rules") || (!hasFlag(args, "--rules") && !hasFlag(args, "--claude"));
-  const writeFlagIdx = args.indexOf("--write");
-  const writeDir = writeFlagIdx >= 0 ? (args[writeFlagIdx + 1] ?? process.cwd()) : null;
-
-  console.log(`${c.bold}${c.cyan}kit analyze${c.reset}`);
-  console.log(`${c.dim}${"─".repeat(50)}${c.reset}\n`);
-
-  const report = await analyzeRepo(process.cwd());
-
-  // Summary line — terse, machine-readable enough to grep
-  const stack = report.stack;
-  console.log(
-    `${c.bold}Detected:${c.reset} ${stack.language}${stack.framework ? ` / ${stack.framework}` : ""}${
-      stack.services.length ? ` + ${stack.services.join(", ")}` : ""
-    }  ${c.dim}(confidence ${(stack.confidence * 100).toFixed(0)}%)${c.reset}`,
-  );
-  if (report.testRunners.length)
-    console.log(`${c.dim}Tests:${c.reset} ${report.testRunners.join(", ")}`);
-  if (report.deployTargets.length)
-    console.log(`${c.dim}Deploy:${c.reset} ${report.deployTargets.join(", ")}`);
-  if (report.databaseClients.length)
-    console.log(`${c.dim}DB clients:${c.reset} ${report.databaseClients.join(", ")}`);
-  if (report.ciFiles.length)
-    console.log(`${c.dim}CI:${c.reset} ${report.ciFiles.length} workflow(s)`);
-  if (report.commitPrefixes.length)
-    console.log(
-      `${c.dim}Commit prefixes:${c.reset} ${report.commitPrefixes
-        .slice(0, 5)
-        .map((p) => `${p.prefix}(${p.count})`)
-        .join(", ")}`,
-    );
-  console.log();
-
-  if (writeDir) {
-    const { writeFile } = await import("node:fs/promises");
-    if (wantClaude) {
-      const path = `${writeDir}/CLAUDE.md.draft`;
-      await writeFile(path, renderClaudeMd(report), "utf-8");
-      console.log(`  ${c.green}✓${c.reset} wrote ${path}`);
-      if (report.hasClaudeMd) {
-        console.log(
-          `    ${c.dim}(existing CLAUDE.md found — review draft before merging)${c.reset}`,
-        );
-      }
-    }
-    if (wantRules) {
-      const path = `${writeDir}/RULES.md.draft`;
-      await writeFile(path, renderRulesMd(report), "utf-8");
-      console.log(`  ${c.green}✓${c.reset} wrote ${path}`);
-      if (report.hasRulesMd) {
-        console.log(
-          `    ${c.dim}(existing RULES.md found — review draft before merging)${c.reset}`,
-        );
-      }
-    }
-    console.log();
-    return true;
-  }
-
-  // No --write: print drafts to stdout for piping / inspection.
-  if (wantClaude) {
-    console.log(`${c.bold}── CLAUDE.md ${"─".repeat(40)}${c.reset}`);
-    process.stdout.write(renderClaudeMd(report));
-    console.log();
-  }
-  if (wantRules) {
-    console.log(`${c.bold}── RULES.md ${"─".repeat(40)}${c.reset}`);
-    process.stdout.write(renderRulesMd(report));
-    console.log();
-  }
-  console.log(
-    `${c.dim}Pipe to a file, or run with ${c.bold}--write .${c.reset}${c.dim} to materialize as CLAUDE.md.draft + RULES.md.draft${c.reset}`,
-  );
-  return true;
-}
 
 export const COMMAND_HELP: Record<string, string> = {
   status: "Adoption checklist — what's set up across kit + the next step for each gap",
@@ -890,49 +233,6 @@ export const COMMAND_HELP: Record<string, string> = {
   help: "Show this help",
 };
 
-async function cmdPkg(): Promise<boolean> {
-  const args = process.argv.slice(3);
-  const input = args[0];
-
-  if (!input || input === "--help" || input === "-h") {
-    console.log(`${c.bold}kit pkg${c.reset} — Install packages with mandatory triage\n`);
-    console.log("Usage:  kit pkg <ecosystem>:<package>[@version]\n");
-    console.log("Ecosystems:");
-    console.log("  npm:express            npm install express");
-    console.log("  npm-g:vercel           npm install -g vercel");
-    console.log("  pnpm:react             pnpm add react");
-    console.log("  pip:requests           pip install requests");
-    console.log("  brew:trivy             brew install trivy");
-    console.log("  docker:postgres        docker pull postgres");
-    console.log("  go:golang.org/x/tools/cmd/goimports@latest");
-    console.log("  cargo:ripgrep          cargo install ripgrep");
-    console.log("");
-    console.log("Examples:");
-    console.log("  kit pkg npm:express@4.18.0");
-    console.log("  kit pkg pip:requests");
-    console.log("  kit pkg docker:stalwartlabs/stalwart");
-    console.log("  kit pkg brew:semgrep");
-    return true;
-  }
-
-  const spec = parsePkgSpec(input);
-  if (!spec) {
-    console.error(`${c.red}Invalid format: ${input}${c.reset}`);
-    console.error("Expected: <ecosystem>:<package> (e.g. npm:express, pip:requests)");
-    return false;
-  }
-
-  console.log(
-    `${c.bold}Installing ${spec.ecosystem}:${spec.name}${spec.version ? `@${spec.version}` : ""}${c.reset}\n`,
-  );
-  console.log(`${c.cyan}Step 1: Triage...${c.reset}`);
-
-  const result = await installPkg(spec);
-  console.log(result.output);
-
-  return result.installed;
-}
-
 function cmdHelp(subcommand?: string): boolean {
   if (subcommand && COMMAND_HELP[subcommand]) {
     console.log(`kit ${subcommand} — ${COMMAND_HELP[subcommand]}`);
@@ -1046,117 +346,6 @@ function cmdHelp(subcommand?: string): boolean {
   console.log(`  ${dim}kit add --list${reset}                 # List all available adapters`);
 
   return true;
-}
-
-async function cmdTeam(): Promise<boolean> {
-  const subcommand = process.argv[3];
-  const args = process.argv.slice(4);
-
-  if (!subcommand || subcommand === "--help" || subcommand === "-h") {
-    console.log(`${c.bold}kit team${c.reset} — manage team members and permissions\n`);
-    console.log(`${c.bold}Usage:${c.reset}`);
-    console.log(`  kit team create <name>                 Create new team`);
-    console.log(`  kit team invite <email> --role=<role>  Invite user to team`);
-    console.log(`  kit team members list                  List team members`);
-    console.log(`  kit team member remove <email>         Remove team member`);
-    console.log(`  kit team audit log [--limit=N]         View audit logs\n`);
-    console.log(`${c.dim}Roles: owner, admin, developer, guest${c.reset}`);
-    return true;
-  }
-
-  try {
-    switch (subcommand) {
-      case "create": {
-        const name = args[0];
-        if (!name) {
-          console.error(`${c.red}Error: team name required${c.reset}`);
-          return false;
-        }
-
-        // Placeholder: would need user context from auth
-        console.log(`${c.yellow}Note: Team creation requires authentication${c.reset}`);
-        console.log(`${c.dim}Team management requires backend service integration${c.reset}`);
-        return false;
-      }
-
-      case "invite": {
-        const email = args[0];
-        const roleArg = args.find((a) => a.startsWith("--role="));
-        const role = roleArg?.split("=")[1] || "developer";
-
-        if (!email) {
-          console.error(`${c.red}Error: email required${c.reset}`);
-          return false;
-        }
-
-        // No backend exists — do NOT print "Invitation sent" for a no-op (that is a
-        // literal false-green, the exact thing kit's thesis condemns). Fail honestly.
-        console.error(
-          `${c.red}Error: kit team invite is not implemented — no team backend is configured${c.reset}`,
-        );
-        console.error(
-          `${c.dim}would invite ${email} as ${role}; team management requires backend service integration${c.reset}`,
-        );
-        return false;
-      }
-
-      case "members": {
-        if (args[0] !== "list") {
-          console.error(`${c.red}Unknown subcommand: ${args[0]}${c.reset}`);
-          return false;
-        }
-
-        console.log(`${c.bold}Team Members${c.reset}\n`);
-        console.log(`${c.dim}(Requires team context — run from project with .kit.toml)${c.reset}`);
-        return false;
-      }
-
-      case "member": {
-        if (args[0] !== "remove") {
-          console.error(`${c.red}Unknown subcommand: ${args[0]}${c.reset}`);
-          return false;
-        }
-
-        const email = args[1];
-        if (!email) {
-          console.error(`${c.red}Error: email required${c.reset}`);
-          return false;
-        }
-
-        // No backend — never claim a removal that did not happen. Fail honestly.
-        console.error(
-          `${c.red}Error: kit team member remove is not implemented — no team backend is configured${c.reset}`,
-        );
-        console.error(
-          `${c.dim}would remove ${email}; team management requires backend service integration${c.reset}`,
-        );
-        return false;
-      }
-
-      case "audit": {
-        if (args[0] !== "log") {
-          console.error(`${c.red}Unknown subcommand: ${args[0]}${c.reset}`);
-          return false;
-        }
-
-        const limitArg = args.find((a) => a.startsWith("--limit="));
-        const limit = limitArg ? parseInt(limitArg.split("=")[1]) : 50;
-
-        console.log(`${c.bold}Audit Log${c.reset}  ${c.dim}(limit: ${limit})${c.reset}\n`);
-        console.log(`${c.dim}No audit events available${c.reset}`);
-        return true;
-      }
-
-      default:
-        console.error(`${c.red}Unknown subcommand: ${subcommand}${c.reset}`);
-        console.error(`Run 'kit team --help' for usage information.`);
-        return false;
-    }
-  } catch (err: unknown) {
-    const error = err as Error;
-    console.error(`${c.red}Error: ${error.message}${c.reset}`);
-    return false;
-  }
 }
 
 /**
@@ -1348,119 +537,6 @@ async function main(): Promise<void> {
       throw err;
     }
   }
-}
-
-/**
- * `kit gate-bash` — PreToolUse install-gate handler. Reads a coding agent's
- * pending-tool-call JSON on stdin ({ tool_name, tool_input: { command } }), and
- * if it is a Bash command that adds an un-triaged package, BLOCKS it by exiting 2
- * (the deny signal for Claude Code / Codex / Amazon Q PreToolUse hooks; exit 1
- * would be a non-blocking error). Allow → exit 0. This is what makes
- * "installs nothing untriaged" hold even in agent auto-mode. Wire it with
- * `kit agent-config --install-gate`. Pure decision lives in install-gate.ts.
- */
-export async function cmdGateBash(): Promise<boolean> {
-  let raw = "";
-  try {
-    const chunks: Buffer[] = [];
-    for await (const chunk of process.stdin) chunks.push(chunk as Buffer);
-    raw = Buffer.concat(chunks).toString("utf8");
-  } catch {
-    return true; // no stdin / read error → do not block
-  }
-  let payload: {
-    tool_name?: string;
-    tool_input?: { command?: unknown };
-    command?: unknown;
-    preToolUse?: { toolName?: string; parameters?: { command?: unknown } };
-  };
-  try {
-    payload = raw ? JSON.parse(raw) : {};
-  } catch {
-    return true; // unparseable hook payload → do not block (avoid breaking the agent)
-  }
-  // Agent-agnostic command extraction (Claude/Codex/Amazon Q/Gemini tool_input.command,
-  // Cursor top-level command, Cline preToolUse.parameters.command) — shared pure helper.
-  const { decideBashGate, extractCommandFromHookPayload } = await import("./install-gate.js");
-  const command = extractCommandFromHookPayload(payload);
-  if (!command) {
-    return true; // no shell command in this tool call → allow
-  }
-  const verdict = await decideBashGate(command);
-  if (verdict.block) {
-    // Cline blocks via a stdout JSON {cancel:true} contract (HookOutputSchema),
-    // NOT exit 2 — so `--format cline` emits that and exits 0; every other agent
-    // uses the exit-2 PreToolUse deny.
-    if (gateFormat() === "cline") {
-      console.log(
-        JSON.stringify({
-          cancel: true,
-          errorMessage: `kit install-gate: ${verdict.reason} — triage first (kit triage …) or install via kit pkg <eco>:<name>`,
-        }),
-      );
-      return true;
-    }
-    const { writeSync } = await import("node:fs");
-    writeSync(
-      2,
-      `kit install-gate: BLOCKED — ${verdict.reason}\nTriage it first: \`kit triage …\`, or install via \`kit pkg <eco>:<name>\`.\n`,
-    );
-    process.exit(2); // PreToolUse deny
-  }
-  return true;
-}
-
-/**
- * PreToolUse hook body for the env-write-gate: block a Write/Edit that puts a
- * plaintext secret into a real `.env*` file, BEFORE it lands. Mirrors gate-bash:
- * fail-open on unparseable payloads (never break the agent), exit-2 deny on block.
- */
-export async function cmdGateEnv(): Promise<boolean> {
-  let raw = "";
-  try {
-    const chunks: Buffer[] = [];
-    for await (const chunk of process.stdin) chunks.push(chunk as Buffer);
-    raw = Buffer.concat(chunks).toString("utf8");
-  } catch {
-    return true; // no stdin / read error → do not block
-  }
-  let payload: unknown;
-  try {
-    payload = raw ? JSON.parse(raw) : {};
-  } catch {
-    return true; // unparseable hook payload → do not block (avoid breaking the agent)
-  }
-  const { extractWriteFromHookPayload, decideEnvWriteGate } = await import("./env-write-gate.js");
-  const write = extractWriteFromHookPayload(payload);
-  if (!write) return true; // not a file write → allow
-  const verdict = decideEnvWriteGate(write.filePath, write.text);
-  if (verdict.block) {
-    if (gateFormat() === "cline") {
-      console.log(
-        JSON.stringify({
-          cancel: true,
-          errorMessage: `kit env-gate: ${verdict.reason} — resolve secrets with \`kit secrets\` (vault-backed) instead of plaintext .env`,
-        }),
-      );
-      return true;
-    }
-    const { writeSync } = await import("node:fs");
-    writeSync(
-      2,
-      `kit env-gate: BLOCKED — ${verdict.reason}\nNever write secrets to .env* in plaintext. Resolve them with \`kit secrets\` (vault-backed), or use a placeholder in .env.example.\n`,
-    );
-    process.exit(2); // PreToolUse deny
-  }
-  return true;
-}
-
-/** The `--format <fmt>` value for gate-bash (`--format cline` | `--format=cline`). */
-function gateFormat(): string {
-  const argv = process.argv;
-  const eq = argv.find((a) => a.startsWith("--format="));
-  if (eq) return eq.slice("--format=".length);
-  const i = argv.indexOf("--format");
-  return i !== -1 && i + 1 < argv.length ? argv[i + 1] : "";
 }
 
 // Single source of truth for kit's top-level command surface. Every key here MUST
